+++ lib/src/settings.dart
@@ -324,23 +324,32 @@ class SettingsTile extends StatelessWidget {
     Widget? trailing,
     double? value,
   })  : trailing = trailing ??
-            // TODO: This is not updated when timeDilation changes outside
             Text(value != null ? value.toString() : timeDilation.toString()),
         onTap = (() => showDialog(
               context: context,
-              builder: (_) => CustomDialog.simple(
-                title: dialogTitle ?? title,
-                child: CustomSlider(
-                  min: 1,
-                  max: 10,
-                  divisions: 9,
-                  value: value ?? timeDilation,
-                  onChanged: (num value) {
-                    timeDilation = value.toDouble();
-                    onChanged?.call(timeDilation);
+              builder: (_) {
+                double tmpValue = value ?? timeDilation;
+                return StatefulBuilder(
+                  builder: (BuildContext context, StateSetter setState) {
+                    return CustomDialog.simple(
+                      title: dialogTitle ?? title,
+                      child: CustomSlider(
+                        min: 1,
+                        max: 10,
+                        divisions: 9,
+                        value: tmpValue,
+                        onChanged: (num value) {
+                          setState(() {
+                            tmpValue = value.toDouble();
+                          });
+                          timeDilation = tmpValue;
+                          onChanged?.call(timeDilation);
+                        },
+                      ),
+                    );
                   },
-                ),
-              ),
+                );
+              },
             )) {
     // set default based on the input value
     if (value != null) {