+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Flutter Widgetz
 
+## [15.2.1]
+
+- Add optional `ScrollController` param to `GroupedListView`.
+- Fix `CustomSlider` label.
+- Fix `SettingsTile` time dilation tile.
+
 ## [15.2.0]
 
 - Introduce `SegmentedSwitch` widget.